organize entry points and update readme
fix

note
