+++ src/test-crawler.ts
@@ -1,0 +1,10 @@
+import { startLocalApi } from "./local-backend-proxy";
+import { crawlRepeatedly } from "./crawl-repeatedly";
+
+const _ = require("dotenv").config();
+
+(async () => {
+  startLocalApi(3000);
+
+  return crawlRepeatedly("dev", "http://localhost:3000/", {});
+})();
